pyproject.toml
[tool.poetry]
name = "cimulator"
packages = [{ include = "cimulator", from = "src" }]
version = "0.1.0"

[tool.poetry.dependencies]
python = "^3.10"
iniconfig = "2.1.0"
packaging = "24.2"
pluggy = "1.5.0"
pyyaml = "6.0.2"
setuptools = "78.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

[tool.poetry.scripts]
cimulator = "cimulator.cli:main"
